Pull out inner class for Dynamic Profile Provider Service.  Use default for metadata key value
